projects/deep-dispatch/requirements.txt
google-genai>=1.0.0
weasyprint>=61.0
markdown>=3.6
python-dotenv>=1.0.0
